+++ src/app/page.tsx
@@ -4,8 +4,8 @@ import { Countdown } from '@/components/Countdown';
 export default function Home() {
   return (
     <>
-      <ExampleSection />
       <Countdown />
+      <ExampleSection />
     </>
   );
 }
