+++ src/components/Countdown/types.ts
@@ -1,0 +1,3 @@
+export interface CountdownProps {
+  className?: string;
+}
