+++ src/data/common.json
@@ -17,5 +17,6 @@
   "contacts": {
     "title": "",
     "email": ""
-  }
+  },
+  "countdown": 24
 }
